requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
diff_gaussian_rasterization==0.0.0
fvcore==0.1.5.post20221221
idna==3.11
imageio==2.35.1
iopath==0.1.10
numpy==1.24.4
opencv-python==4.12.0.88
pillow==10.4.0
pip==24.2
plyfile==1.0.3
portalocker==3.0.0
pytorch3d==0.3.0
PyYAML==6.0.3
requests==2.32.4
setuptools==75.1.0
simple_knn==0.0.0
tabulate==0.9.0
termcolor==2.4.0
tinycudann==2.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
wheel==0.44.0
yacs==0.1.8
